Tools Extra Abilities + Fixes
TOOLS EXTRA ABILITIES:

+ Added the ability to place a torch upon right click while holding the
Tenebrae Pickaxe (If any torches are available on the player’s
inventory)

+ Added the ability of being able to destroy
dirt/gravel/sand/ash/snow/grass with the Tenebrae Pickaxe if the player
has a Tenebrae Shovel on the inventory.

FIXES:

+ New death sequence for the Empowered Tenebrae Golem, and he will give
now more experience when killed.

· Fixed and changed the Tenebrae Infuser recipe.

· Entities should now have the language localization.

· Modified the statistics of the Pyro Knight, removed AI’s, changed the
weapon that it was wielding from a stick to a Tenebrae Sword, and added
a new ability (The Pyro Knight will set the block he is standing on
fire when not in water).

· Minor fixes to the Tenebrae Golem and  Empowered Tenebrae Golem model
file.